backend/requirements.txt
# fastapi==0.115.0
# uvicorn==0.30.6
# python-multipart==0.0.9
# requests==2.32.3
# streamlit-audiorecorder==0.5.0  # for frontend audio


# google-cloud-aiplatform==1.60.0
# google-cloud-firestore==2.20.0
# google-cloud-storage==2.17.0
# google-cloud-speech==2.26.0


fastapi
uvicorn
python-multipart
google-cloud-firestore
google-cloud-speech
google-cloud-storage
google-cloud-aiplatform

